feat: add main test for backend
