+++ package.json
@@ -11,7 +11,8 @@
     "express": "~4.18.1",
     "graphql": "^16.9.0",
     "react": "18.3.1",
-    "react-dom": "18.3.1"
+    "react-dom": "18.3.1",
+    "supertest": "^7.0.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.8.0",
